requirements.txt
Appium-Python-Client==1.0.1
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.3
requests==2.23.0
selenium==3.141.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.3